Moves Calibrator to calibration package.
